feat(ListDiscussions): restructure component
